third_party/WebKit/LayoutTests/typedcssom/stylevalue-subclasses/cssScale-expected.txt

This is a testharness.js-based test.
PASS Constructing a CSSScale with an angle CSSUnitValue for the coordinates throws a TypeError
FAIL Constructing a CSSScale with a CSSMathValue that doesn't match <number> for the coordinates throws a TypeError assert_throws: function "() => new CSSScale(coord, 0)" did not throw
PASS Updating CSSScale.x to an angle CSSUnitValue throws a TypeError
PASS Updating CSSScale.x to a CSSMathValue that doesn't match <number> throws a TypeError
PASS Updating CSSScale.y to an angle CSSUnitValue throws a TypeError
PASS Updating CSSScale.y to a CSSMathValue that doesn't match <number> throws a TypeError
PASS Updating CSSScale.z to an angle CSSUnitValue throws a TypeError
PASS Updating CSSScale.z to a CSSMathValue that doesn't match <number> throws a TypeError
FAIL CSSScale can be constructed from two number coordinates assert_equals: expected "CSSUnitValue" but got "Number"
FAIL CSSScale can be constructed from three number coordinates assert_equals: expected "CSSUnitValue" but got "Number"
FAIL CSSScale can be constructed from CSSMathValue coordinates assert_equals: expected "CSSMathSum" but got "Number"
FAIL CSSScale.x can be updated to a number assert_equals: expected (object) object "3.14" but got (number) 3.14
FAIL CSSScale.x can be updated to a numberish assert_equals: expected "CSSUnitValue" but got "Number"
FAIL CSSScale.x can be updated to a CSSMathValue assert_equals: expected "CSSMathSum" but got "Number"
FAIL CSSScale.y can be updated to a number assert_equals: expected (object) object "3.14" but got (number) 3.14
FAIL CSSScale.y can be updated to a numberish assert_equals: expected "CSSUnitValue" but got "Number"
FAIL CSSScale.y can be updated to a CSSMathValue assert_equals: expected "CSSMathSum" but got "Number"
FAIL CSSScale.z can be updated to a number assert_equals: expected (object) object "3.14" but got (number) 3.14
FAIL CSSScale.z can be updated to a numberish assert_equals: expected "CSSUnitValue" but got "Number"
FAIL CSSScale.z can be updated to a CSSMathValue assert_equals: expected "CSSMathSum" but got "Number"
PASS Modifying CSSScale.is2D can be updated to true or false
Harness: the test ran to completion.
